+++ helpers/helper.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require_relative 'query_utils'
+
 helpers do
   def h(text)
     Rack::Utils.escape_html(text)
@@ -12,16 +14,6 @@ helpers do
   def connection
     @connection ||= PG.connect(dbname: 'memo_app')
   end
-
-  def find_memo(id)
-    sql = 'SELECT * FROM memos WHERE id = $1;'
-    connection.prepare('find', sql)
-    params = [id]
-
-    connection.exec_prepared('find', params) do |result|
-      result.each.with_object([]) do |row, array|
-        array << row
-      end
-    end
-  end
 end
+
+helpers QueryUtils
